+++ framework/core/Assets/Assets.php
@@ -145,6 +145,15 @@ class Assets
 						);
 					}
 				}
+
+				/**
+				 * Scripts localization
+				 *
+				 * @since 1.11.4
+				 */
+				foreach ( config( 'app.assets.scripts.localize', [] ) as $localize ) {
+					wp_localize_script( ...$localize );
+				}
 			}
 		);
 	}