+++ framework/core/Models/Taxonomy.php
@@ -69,8 +69,8 @@ use Brocooly\Support\Builders\TaxonomyQueryBuilder;
  * @method static self trashed()
  * @method static self future()
  *
- * @method static array all()
- * @method static array get()
+ * @method static \Timber\PostQuery all()
+ * @method static \Timber\PostQuery get()
  * @method static self query( array $query )
  * @method static self with( array $postsIn )
  * @method static self except( array $postsIn )