+++ web/wp-config.php
@@ -49,7 +49,7 @@ if ( version_compare( $brocooly_min_php_version, phpversion(), '>=' ) ) {
 		/* translators: 1 - minimum required PHP version, 2 - current PHP version. */
 		sprintf(
 			/* html */
-			'<h1>Brocooly Framework requires PHP version %1$s or greater!</h1><p>Invalid PHP version! Please update it. Your current version is: <strong>%2$s</strong></p>',
+			'<h1>Brocket Framework requires PHP version %1$s or greater!</h1><p>Invalid PHP version! Please update it. Your current version is: <strong>%2$s</strong></p>',
 			esc_html( $brocooly_min_php_version ),
 			esc_html( phpversion() ),
 		),